Reduce the use of compute_call_sites()
The lowerForalls sub-pass of the compiler makes extensive use of
flattenNestedFunction(FnSymbol*) to flatten newly-created clones of
task functions. This results in a significant portion of compilation
time being spent in compute_call_sites().

This is because flattenNestedFunction(FnSymbol*) invokes
flattenNestedFunctions(Vec), which invokes compute_call_sites(), which
sets up the calledBy vector on every live FnSymbol. Which, among others,
includes allocation of calledBy vectors and possibly deallocation of their
old contents, and traversing the dispatch children of dynamically-dispatched
FnSymbols. Given that flattenNestedFunction(FnSymbol*) needs to look
only at a handful of other functions, if at all, and none of those
participate in dynamic dispatch, most of compute_call_sites() work
is unused upon a flattenNestedFunction(FnSymbol*).

This change adds a `fastCCS` argument to flattenNestedFunctions(Vec).
When it is invoked from flattenNestedFunction(FnSymbol*), this
argument is enabled. This causes flattenNestedFunctions(Vec)
to avoid compute_call_sites(), instead computing `calledBy` only
for those functions that will be looked at. I added an assertion
that these functions do not participate in dynamic dispatch, because
otherwise `calledBy` computation may be incorrect.