Rework transmit context handling to improve comm/compute overlap.
The documentation indicates that we need to be more aggressive about
consuming CQ entries when we get -EAGAIN initiating a transaction.  And
we also need to be able to overlap communication and computation more
than we are now, certainly for situations where we don't have fixed
threads (and Tx contexts for each).  To that end, rework the transmit
context management, implement a reallocator for Tx contexts, and release
Tx contexts while waiting for completions.

While here, relieve readCQ()'s callers of the need to check for -EAGAIN
by consuming that value within the routine itself.