Clean up provider-specific support; allow for "special error" reporting.
Recent (even current) experience indicates we're going to need a fair
amount of provider-specific logic in the ofi comm layer going forward.
Here, reorganize the provider support so that sort of thing will be
easier to implement in the future.

As an outgrowth of the above, add a "special" error handler to print
more-informative messages when certain unrecoverable errors occur and we
have some idea what the user could do to avoid them.  Add the first such
error (open-file count overrun with tcp) into that.