Make the runtime utilize CHPL_LAUNCH_MASTERIP to set GASNET_MASTERIP when unset
We'd like to use the value that GASNET_MASTERIP normally gets set to during
multilocale interop, but it would be odd to rely on a gasnet environment
variable in the situations where gasnet isn't used (such as for --no-local
compilations, or when only CHPL_LAUNCHER is set but not CHPL_COMM, both of which
we intend to support eventually).  So instead, add this use of an environment
variable that is Chapel-specific and utilize it to set the equivalent gasnet
environment variable.

Checked behavior when:
- GASNET_MASTERIP is already in the environment but CHPL_LAUNCH_MASTERIP is not
(behavior: Chapel executable compiles and runs to completion)
- neither GASNET_MASTERIP nor CHPL_LAUNCH_MASTERIP is in the environment
(behavior: Chapel compiles, but the executable fails to communicate, as
expected)
- only CHPL_LAUNCH_MASTERIP is set (behavior: Chapel executable compiles and
runs to completion)
- GASNET_MASTERIP is set to a valid value and CHPL_LAUNCH_MASTERIP is set to
something bogus that would not work on its own (behavior: CHPL_LAUNCH_MASTERIP's
value does not override the GASNET_MASTERIP setting, and so the program compiles
and runs to completion)