Fix check-http-setpopt test for overload set checking
Follow-on to PR #13442.
Note that in this case the overload set error had to do with
two `extern` declarations of `curl_easy_getinfo`. I suspect that
multiple overloads for an `extern` function is harmless, because
it is not up to Chapel at all to decide which overload to call
(since C functions do not have overloading).