Fix issue where GASNET_MASTERIP was not getting set early enough when required
Greg realized that we needed to set the variable during prelaunch.  This change
add new files comm-gasnet-launch.c and comm-gasnet-launch.h, similar to
comm-ugni-launch.h and comm-ugni-launch.c.  It then restructures
comm/gasnet/Makefile.share to build the .c file in the same situations as the
ugni one.

With this change, CHPL_LAUNCH_MASTERIP seems to appropriately set
GASNET_MASTERIP when GASNET_MASTERIP is required!