Add string.toByte() and string.toCodepoint() (#13553)
Add string.toByte() and string.toCodepoint() (#13553)

[reviewed by @mppf and @cassella ]

The deprecated ascii() function was replaced by string.byte(1), which provided extra functionality by varying its argument, but looked a little heavy-weight for the purposes for which ascii() was used. This change adds string.toByte() as a better direct replacement for ascii() (while still retaining the more general string.byte()). This change also adds string.toCodepoint() as the codepoint analog.

The key difference between ascii() and string.toByte() is that the latter will issue an error message if the string is not exactly one byte long. Likewise, string.toCodepoint() will issue an error message if the string is not exactly one codepoint long.

Some existing code needed a param version of ascii(), so a param version of string.toByte() has been created as well. There has not yet been a need for a param string.toCodepoint(), but that could be added in the future. (See Cray/chapel-private#348 .)

Existing Chapel code that had been changed from using ascii() to string.byte(1) is changed here to use string.toByte() wherever it was operating on a string that was required to be one byte long.

Some tests have been added to fill in gaps.

At some point, the String module should be migrated to use HaltWrappers. However, HaltWrappers are not yet able to handle some of the error messages in String, so the migration has not occurred yet. Consequently, this change stays consistent with the rest of String by using halt() and leaves the migration to a future PR. (See Cray/chapel-private#349 .)

Note: modules/packages/LAPACK.chpl is a very large diff. The only change to that file is to change all occurrences of .byte(1) to .toByte(). These were all former uses of ascii().

Passed full local and full GASNet testing on linux64.
Passed Cray XC testing (particularly LAPACK) as well as --fast.
Retested after review.

Closes #13167
Closes Cray/chapel-private#80