In instantiateFunction, leave defaultExpr block
classes/initializers/compilerGenerated/generics/syncFieldTypeOnlyGeneric2.chpl
was running into a problem where the block from a formal's defaultExpr
(in the initializer) was used as the instantiation point for a generic
type. instantiateFunction was removing such blocks in the process of
replacing their contents. The result was an internal error about missing
a visibility block when resolving something to do with the type, because
the instantiated type's instantiation point referred to the removed
block.

This commit just adjusts the code replacing the defaultExpr to leave
the outer BlockStmt there if it existed. Since we don't expect nested
BlockStmts (at least from users) in defaultExprs, that should be
sufficient for the time being.