Nil checking for owned, shared types
This generates an error when default-initializing a variable
of a **non-nilable** owned or shared class type:
```chpl
var x: owned MyClass;
var y: shared MyClass;
```

I also adjusted most tests where this new error would be triggered
when run with `--no-legacy-nilable-classes`.
I added a test to lock in the error message.

The implementation refactors and reuses compiler code that issues error
for similar cases with borrowed/unmanaged types. A special case for an
owned/shared type within a tuple type is needed so that the line number
in the error message points into the user code, rather than ChapelTuple.chpl.

I considered implementing this error by extending the default initializer
for _owned/_shared in the module code. I chose to do it in the compiler
because we can generate nicer error message in that case.

Things that came up and not addressed in this PR:

* Checking is not done on the element type of a default-initialized array.
This is the same as when the element type is borrowed/unmanaged.

* Weird compiler behavior and/or errors:
```
classes/delete-free/lifetimes/borrow-escapes.chpl  // if we remove '!' in c!.borrow()
expressions/ferguson/make-wide.chpl
studies/bale/toposort/toposort.chpl
release/examples/spec/Classes/owned-assignment.chpl
```

* Missed checking in this case:
```chpl
// classes/ferguson/delete-free/owned-shared-fields2.chpl
record R5 {
  var fo:owned MyClass = nil;
  var fs:shared MyClass = nil;
}
var r = new R5();
```

* Missed checking in this case:
```chpl
// optimizations/widepointers/refOverride.chpl
  var printer: owned SuperPrinter?;
  printer = new owned SubPrinter();
  printer.print(data);
```

* Spurious warnings "undecorated class type MyGenericClass is unstable"
```chpl
// compflags/ferguson/unstable-class.chpl
var eg:owned MyGenericClass(int)?;
var fg:shared MyGenericClass(int)?;
```

* It seems desirable for the user to request that the runtime check
performed upon a '!' be skipped upon `--fast`. For example, if I want
to get maximum performance out of:

```
users/npadmana/twopt/do_smu_aos
users/npadmana/twopt/do_smu_soa
```

there seems to be an invariant that if a node is not "isLeaf", then
its both 'left' and 'right' fields are non-nil, so I would like to skip
the runtime check for '!' in arguments to the recursive calls to
TreeAccumulate().

* There is a surprisingly high number of cases where the default initialization
is immediately followed by an assignment, after which the variable is assumed
to be non-nilable. A few of those cases are perhaps intentional for testing
and/or convertable to non-default initialization in a straightforward
way. However, there are a couple of more interesting cases:

```chpl
  // users/npadmana/twopt/do_smu_soa.chpl
  var pp1, pp2 : owned Particle3D;
  if isPerf {
    pp1 = new owned Particle3D(nParticles, true);
    pp2 = new owned Particle3D(nParticles, true);
  } else {
    pp1 = readFile(fn1);
    pp2 = readFile(fn2);
    if !isTest {
      writef("Read in %i lines from file %s \n", pp1.npart, fn1);
      writef("Read in %i lines from file %s \n", pp2.npart, fn2);
    }
  }
```
```chpl
  // studies/bale/toposort/toposort.chpl
  var topoResult : shared TopoSortResult(D.idxType);

  select implementation {
    when ToposortImplementation.Serial {
       if !silentMode then writeln("Converting to CSC domain");

      var dmappedPermutedSparseD : sparse subdomain(D) dmapped CS(compressRows=false);
      dmappedPermutedSparseD.bulkAdd( permutedSparseUpperTriangularIndexList );

      if !silentMode then writeln("Toposorting permuted upper triangluar domain using Serial implementation.");
      topoResult = toposortSerial( dmappedPermutedSparseD );
    }
    ... other cases are similar ...
    otherwise {
      writeln( "Unknown implementation: ", implementation );
      exit(-1);
    }
  }
```

also the former `var r: owned Runner;` in `arrays/diten/timeVectorArray.chpl`