Take advantage of known UTF-8 encoding to speed up strings (#13580)
Take advantage of known UTF-8 encoding to speed up strings (#13580)

[reviewed by @mppf ]

This change speeds up the substring operation by more than 3x to address string performance regressions discussed in #13130 .

The String module was originally updated in a character-set-agnostic way. Now that we know the character set is guaranteed to be UTF-8, we can take advantage of that fact not just to speed up individual operations, but to choose faster algorithms. For example, there are now some cases where a multibyte character does not have to be decoded, where it was before.

The 3x performance benefit was observed in the substring portion of test/types/string/psahabu/perf/substring.chpl.

Passes full local and GASNet testing.