In cleanupAfterRemoves in resolution, NULL removed substitutions
PR #13555 updated the way that substitutions
are computed and after PR #13447 it is possible that some of the
substitutions for type arguments are in fact generic types.
In that event, the types are removed from the tree by resolution
cleanup functions. However, they were not being removed from
fn->substitutions, causing use-after-free errors in
convertClassTypes.

This PR adjusts cleanupAfterRemoves() to put a NULL in any
substitution value that has been removed from the tree.