Don't prepend /usr/lib64 to LD_LIBRARY_PATH any more.
We used to have to prepend /usr/lib64 to LD_LIBRARY_PATH, because Shasta
slurm was running with --mpi=pmi2 by default and we needed the slurm PMI
libraries in that directory.  But Shasta slurm recently began running
with --mpi=cray-pmi by default, so we need the Cray PMI libraries.
Removing the prepended /usr/lib64 causes us to find those instead of the
slurm ones.

Note that a module built with this change will do the wrong thing if it
is used in an "old" Shasta environment built before about 2019-07-25,
but those are rapidly becoming rare.