Adjust borrow-escapes.chpl for nil checkng
In #13574 I modified this test to pass nil checks.
As a result, the code started to invoke `!` on an empty field.
Now, instead, always initialize the field so it is non-nilable.
Also, remove a TODO comment that is now obsolete.
Note that these nilability issues are orthogonal to the purpose
of the test, which is to test certain lifetime errors.

Discussed with @mppf.