Fix memory leaks in test-radix-bucketizer.chpl
Resolves #13569

When setting up the AST for a task reduce intent, I was using
PRIM_ASSIGN instead of calling the `=` function (which is a really bad
idea when used on an array btw) to store the result of globalOp.generate()
into the outer variable.  Also, the task-private shadow variables were not
getting autoDestroy-ed. This fixes these bugs.