Stop trying to load a cmake module for the XC-x86 module build
When we built the module on a CLE5 machine, we needed to load a newer
cmake and python in order to build llvm. That's not needed on CLE6 or
newer, so stop loading them. Note that the arm module already stopped
doing this since it was running on a newer machine. Additionally remove
the code from the shasta module, it isn't building llvm support yet, but
when it does this code won't be needed.