Add a future for the German sharp S upper/lower case anomaly (#13649)
Add a future for the German sharp S upper/lower case anomaly (#13649)

[new future, not reviewed]

As discussed in #13648 , the Unicode German sharp S takes up a different number of bytes for its upper vs. lower case versions. This change adds a future to track the string methods' assumption that the upper and lower case versions would encode to the same number of bytes.

Tested on Cygwin, Darwin, and Linux (including Cray XC systems).