Add an error message for ranges over enums with just one value
At present, the range, domain, and array code relies on being able to
create empty ranges of the form 1..0 or an equivalent.  For enums,
this is done by creating a range that goes from the second enum to the
first enum.  Issue #13650 points out that this was causing a confusing
execution time error message in practice for enums that only defined a
single symbol.  Here, I'm inserting a compile-time error message to
more proactively handle such cases as a placeholder until we either
determine what to do about such cases (e.g., in issue #5165, Louis
proposes that we could have an `empty range` bit rather than, or in
addition to, relying on low > high values.  We could also decide more
formally not to support such cases.