Don't trash the provider name!
I was using strtok() directly on the provider name in the domain
descriptor.  This was idiotic.  strtok() modifies its argument, in this
case replacing all the ';' characters with '\0' characters.  Thus after
the first call to isInProviderName() any utility providers would be
hidden from anyone (including isInProviderName() itself) who looked at
the provider name.  It's a bit surprising anything worked at all.

With this change one can force use of the 'verbs' provider without
causing a hang during initialization, which was what was happening
before.  However, we're still seeing half the stream performance we had
with 'verbs' before this PR.