Restore stream performance.
Work on this branch had caused a 2x performance drop for 16-node stream
with the "verbs" provider.  Here, restore that performance by calling
sched_yield() rather than local_yield() while waiting for the initial
network response to an AM request for an executeOn.

The performance drop was about the right magnitude to be due to a NUMA
mismatch, where the timed kernel iterations assigned to some task were
associated with memory on one NUMA node but the task ended up running on
the other NUMA node.  (The timing in stream is done in such a way that
the reported performance is determined by the slowest worker task.)  If
the local_yield() could cause a Qthreads worker pthread switch, that
could explain the performance regression.