Fix the Unicode sharp S bug (#13662)
Fix the Unicode sharp S bug (#13662)

[reviewed by @e-kayrakli ]

Upper/lower case-changing string methods assume that the upper case and lower case version of a character encode to the same number of UTF-8 bytes. The German sharp S character violates that assumption, as detailed in #13648 .

The MacOS towlower() and towupper() functions resolve the issue by refusing to change the case of the character in question. This change adopts that solution for Chapel on all platforms. When a case change is attempted and the new character would encode to a different number of bytes than the old one, the case change is not performed.

Passes full local and GASNet testing on linux64.

Closes #13648