Various final (I hope) tidying up.
Rename AM and transaction (network) 'done' flags for clarity, and
make the transaction-done flags atomic boolean instead of the less
trustworthy volatile int.

While here, also don't make ensure_progress() calls out of checkTxCQ()
if we're the AM handler.  We're already doing such calls explicitly from
the main AM handler loop, but with a throttle because we don't want the
AM handler doing too much of that sort of thing.