In comm=ofi, don't yield during the network wait for executeOnNB AMs.
While working on #13612 we had noticed that if we did a local_yield()
instead of a sched_yield() while waiting for the network transaction to
complete after initiating an AM, the performance of global stream was
cut in half.  It turns out that this was due to the comm=ofi equivalent
to the problem fixed in comm=ugni by #7915:

- Locale 0 does a local task-spawn to queue the tasks for the inner
  coforall across the PUs.
- Locale 0 does a remote task-spawn to locale 1 for the coforall-on
  across nodes.
- Locale 0 yields while waiting for the network response from locale 1.
- The local tasks queued on locale 0 in the first step all run.
- Locale 0 finishes the kernel and resumes spawning remote tasks for the
  coforall-on.
- The remaining locales queue their local tasks and finish the kernel.

This effectively doubles the execution time, because we don't fire the
remote tasks for nodes 2 through numNodes-1 until after node 0 has done
the whole kernel.

Here, implement essentially the same solution for comm=ofi that we did
for comm=ugni in PR #7915, namely: don't yield while waiting for the
network response for an executeOnNB.  This also lets us do local_yield()
instead of using the stopgap sched_yield() put in with #13612 to solve
the stream performance regression, improving the likelihood of getting
good compute/communicate overlap in overloaded configurations in the
future.

While we're at it, when we do yield while waiting for the network
response for an AM, only yield every 64 times around the completion
check loop.  This should reduce overheads due to talking to network
interfaces without unduly increasing the latency of noticing
completions.