Implementation of the `set` type, to replace "array-as-set" operations (#13546)
Initial implementation of the `set` type (#13546)

This PR provides the initial implementation of the `set` type, intended
to act as a suitable replacement for "domain-as-set" operations.

A set is a collection of unique elements. Sets are unordered and
unindexed. Attempting to add a duplicate item to a set has no effect.
Sets may be iterated over, but do not guarantee a consistent iteration
order.

Almost every `set` method has some testing coverage, except for the
`writeThis` method, for which I could not think of a suitable test.

The `readWriteThis` method has not been added yet - I will leave that
as a future endeavor.

The destructor test is currently a future, due to associative domains
default initializing and destroying _all_ their backing store (this
includes unused slots).

Thanks to @LouisJenkinsCS, @e-kayrakli for review.