Generalize my error message that was previously specific to zippering
In looking at this PR, Elliot pointed out that any case that doesn't
inline can't be handled similar to the zippered case that I'd
focused on before, so here I've generalized the error message to
talk more in terms of inlining iterators and less specifically about
zippering.  I also added a test of throwing --no-inline-iterators
to show a different way that this error message could come up.