Have #0 operator preserve a range's given bounds
Prior to this change, applying #0 to a range would always result
in the range 1..0 coming back.  This meant that, for example,
`42..#0` would return `1..0` where arguably the answer `42..41`
would be better since it preserves the given bound (and similarly,
`10..20#0` should give `10..9`.  Here, I reimplement the #0
case to preserve the appropriate bound of the original range.

One complicated case is the one where computing the "next element"
in the direction that results in an empty range wraps around the
legal range of idxType.  For example, (0:uint)..#0 would like
to return 0..-1 but since -1 can't be represented, it's not clear
what should happen.  In discussion on issue #5165, we decided to
apply normal integral math and have the normal behavior occur
(which suggests uint values will wrap around and int values can't
be depended upon).