Drop llvm nightly testing from nodepara 8 to 2
llvm testing used to run on a 16 core machine, and we ran with nodepara
8 to speed up testing (I would guess there were other jobs running on
the same machine so we needed it to finish quickly.) The llvm job was
moved to a much smaller 4-core VM, so nodepara 8 is way too much. Just
drop to nodepara 2, which should be plenty fast and avoid some timeouts
we've been seeing.