Fix a new test that I didn't test enough
Sloppy sloppy sloppy...  In reviewing PR #13595, Elliot asked a
question that caused me to add a new test case, but I only tested
it for CHPL_COMM=gasnet, not none (though the original PR had been
tested heavily in both configurations).  This fixes it by forking
it from a compopt to a separate test and skipping it for comm=none
(it's not so interesting in that case).  Sorry for the mistake.