Nilability checking improvements #2 (#13677) by vass
* Implement some previously-missing checks.
* Improve the error messages in some cases.
* Minor fixes in two existing tests.
* Add testing of all combinations of nilability and memory management
  of lhs and rhs when initializing a variable this way:

        var lhs: <lhs type> = rhs;

The tests are added under:

    test/classes/errors/nilability-init-var/

Some improvements are implemented as additional `init=` methods or
otherwise in the modules. Other improvements are from compiler changes.

In some cases, the following code:

    var lhs: <lhs type> = rhs;
    compilerError("something");

results in compilerError() firing, whereas without that call
the compiler would report an error on the `var lhs` line.
This is because the `var lhs` error would be reported in
`insertAndResolveCasts(fn)`, which is called after
`resolveBlockStmt(fn->body)` completes successfully.

Some of the compiler changes in this PR make the compiler still report
the `var lhs` error even in the presence of a subsequent compilerError().

The all-combinations tests were created by running this script:

    classes/errors/nilability-generate.chpl

then manually adjusting two files to comment out compilerError(), see below.
I set to test them with --no-codegen, to save testing time. This is a
consideration because 28 of them compile without errors.

FUTURE WORK

* Remove the error upon `var lhs: <shared?> = <owned!>`, which is legal.
  See `init-var-oknil-shared-from-nonnil-owned.future`

* The following cases would still fire compilerError() despite an error
  in the preceding `var lhs` line. I commented out compilerError() in those
  tests to trigger the desired error messages. We want to see those messages
  ahead of compilerError().

    init-var-nonnil-borrowed-from-oknil-nil.chpl
    init-var-nonnil-unmanaged-from-oknil-nil.chpl

* In #11118 we preferred the error message pattern `cannot assign to <lhs>  from <rhs>`.
  There are a few places in the modules and in the compiler that
  either switch the order or say `assign <lhs> to <rhs>`.
  It would be nice to make those consistent with #11118.

Reviewed by / with suggestions from:  @mppf, @lydia-duncan