Stop ignoring test-venv errors
The test-venv rule invokes the third-party-test-venv rule which
ignores errors.  For nightly testing, this means that when the
`make test-venv` command works, we postpone issuing an error
until we try to run testing and fail to do so.  Here, I'm
removing the `-` that ignores errors on this rule and its
sibling so that we get errors closer to the point of failure.