Nilability checking improvements #3
This is a continuation of #13677, adding testing for the following patterns
in the following subdirectories of test/classes/errors/

    // default field initialization  --  nilability-init-field-dflt/
    record MyRecord {
      var lhs: <lhs type> = rhs;
    }

    // field initialization in init()  --  nilability-init-field-arg/
    record MyRecord {
      var lhs: <lhs type>;
      proc init(in rhs) { lhs = rhs; }
    }

    // assignment  --  nilability-assign/
    var lhs...;
    var rhs...;
    lhs = rhs;

DETAILS

The check added in `proc =(ref lhs:_shared, in rhs:owned)`
is independent of chpl_legacyNilClasses.
This is because currently we need to issue a compilerError() in either case.
When chpl_legacyNilClasses is on, the erroneous scenario would lead to an
"unresolved" error if the compilerError() were not there.

The change in FnSymbol.cpp comes from my observing a function
`init_default_<field name>()` with FLAG_METHOD and without `fn->_this`,
failing the now-removed assert.

This PR extends test/classes/errors/nilability-generate.chpl
and adds the tests that it generates (that were not already added by #13677) :

    test/classes/errors/nilability-init-field-arg/
    test/classes/errors/nilability-init-field-dflt/
    test/classes/errors/nilability-assign/

In `proc init()` in nilability-init-field-arg/, the where-clause is needed
so that the init() is considered for copy-initialization.

This PR also adjusts these tests to match the existing .good:

    test/classes/ferguson/delete-free/owned-shared-fields2.chpl
    test/classes/ferguson/delete-free/owned-shared-fields3.chpl

The errors that would arise in those tests without the adjustment are tested
in nilability-init-field-dflt/.