Handle coercions before default arguments
This commit adjusts wrapAndCleanUpActuals to process coercions and `in`
intent arguments before handling default arguments. This resolves the
problem in issue #10104. The problem there is that a coercion was used
to reach the first argument, but the default expression for the second
argument was finding the type of the actual (without the coercion) and
then the compiler got confused.