mason: Changed /test/mason/mason-test-exit/exitCodeTest from future to skipif (#13712)
Issue: Mason is not built in nightly testing, so tests that rely on mason being built are currently classified as futures. This causes noise for developers when they test in an environment with mason built.

Fix: Changed ExitCodeTest to a skipif test, that checks to see if mason is built, if it is the test will run, if not then the test will be skipped.