Deprecate "array-as-vector" methods and adjust failing tests (#13530)
Deprecate "array-as-vector" methods and adjust failing tests (#13530)

This PR replaces _most_ uses of the following array methods:

- push_front
- push_back
- pop_front
- pop_back
- insert
- remove
- clear

With use of the new `list` type where possible. It rewrites several
packages (Mason, UnitTest, LinearAlgebra, TOML) to make use of the
list type as well. Not all of the conversions are pretty - the list
type currently lacks some routines that would result in more elegant
code.

The important thing here is that insofar as possible - when "array-as
vec" deprecation warnings are turned on, the only tests that will
break are those that need to be _removed_, as they are no longer
relevant in a 1.20 world.

Deprecation warnings are off by default, but can be turned on by
setting the config param "showArrayAsVecWarnings" to true at compile
time.

Deprecation warnings will be turned on by default at around release
time.

When deprecation warnings are turned on, any remaining tests with
vector ops in them will have to either:

- Be removed
- Moved to test/deprecated
- Have their .good files updated to contain the warnings

---

Testing:

- [x] `ALL` on `linux64` when `CHPL_COMM=none`
- [x] `ALL` on `linux64` when `CHPL_COMM=gasnet`

---

Reviewed by more people than I can shake a stick at:

@benharsh, @lydia-duncan, @mppf, @e-kayrakli, @bradcray, @daviditen,
@vasslitvinov, @ben-albrecht