Update implementation of anonymous counted range optimization for #0
This fixes the anonymous counted range optimization for the case of
applying #0 to make it match the module code when the optimization
does not fire.  The difference in behavior was revealed by running
--baseline on anonymousCountedRange.chpl.

In more detail, PR #13673, the definition of #0 was changed to
preserve bounds rather than returning 1..0 as it did previously.
However, in that PR, I failed to note that the anonymous range
optimization also special-cased #0 to map it to 1..0 for
consistency.  Here, I change its implementation to use lo..lo-1
(since it currently only supports unstrided ranges, otherwise it
should use lo..lo-stride) instead, making the behavior consistent
with the normal range module code.

I also updated the anonymousCountedRange test to avoid applying
Then I have it try it as a final measure to make sure that the
error message is generated consistently whether run with or without
--baseline