Retire 'jacobi-unoptimized' test
This removes the jacobi-unoptimized.chpl test, which is identical to
the jacobi.chpl test in the same directory except that it is compiled
with --baseline.  We run the test suite nightly with the --baseline
flag, so running this one test with --baseline separately seems a bit
arbitrary.

The motivation for retiring this test at this time is that for LLVM
testing, it revealed an LLVM+baseline-specific bug.  That bug is
worthy of being addressed, but it seems it should be related to
getting LLVM+baseline to the point that it can across the regression
suite cleanly rather than keeping LLVM testing clean in the default
testing mode.