Second attempt to fix optimized loops over counted ranges (lo..#count)
The code I added in PR #13741 turned out to have a problem when
the low bound of a range was min(int) even when the count was not
zero (no wraparound occurred) because we generated expressions
that seemed to wrap around in the branch of the conditional not
taken, but the C compiler couldn't know that it would not be taken.
Here, I'm avoiding that issue by doing the math on uints (where
wraparound is legal) and then casting back to the original index type.

This problem was exhibited by test/distributions/diten/Dimensional.chpl
though I've added a more direct test that locks the problematic case
in here.