Add shim/overloads of methods converting array to list (#13745)
Add shim/overloads of methods converting array to list (#13745)

This PR fixes compiler errors that prevented mason from building. The
errors were introduced from the merge of #13530. Shamefully, I never
thought to build mason itself before merging that PR.

I fixed the compiler errors by adding overloads of the unresolved
methods that convert an array to a list before passing it on, as a
temporary fix.

Testing:

- [x] `cd tools/mason && make` on `linux64` when `COMM=none`
- [x] `test/mason` on `linux64` when `COMM=none`

Trivial and not reviewed.