Fix LCALS performance test failure caused by #13530 (#13757)
Fix LCALS performance test failure caused by #13530 (#13757)

This PR fixes a test failure caused by the merge of #13530.

The LCALS module relies on a vector wrapper that uses 0-based
indexing. When originally converting this code to use `list`, I failed
to notice this. Thus recent test runs using the new code would fire
an assert due to an out of bounds index.

This is particularly painful for the LCALS performance benchmark
because it has a 30 minute plus execution time.

Converting from 0 to 1 based indexing in `vector.this` makes list
happy and fixes the test failure.

Testing:

- [x] `test/release/examples/benchmarks/lcals/LCALSMain.chpl` with
      `-performance` on `linux64` with `COMM=none`
- [x] `test/release/examples/benchmarks/lcals` on `linux64` with
      `COMM=none`

Trivial and not reviewed.