Explicitly set the craype version
We were seeing a hang while building (or rather, failing to build) gmp with
the cray compiler.  This seems to be due to a small bug introduced in
craype/2.1.6.11, at least with the older versions of cce that we've pinned to.
I think we can stop performing this fix when we update compiler versions
again, since I didn't see the issue with those.  I left some comments to that
effect in the file.