Allow receives on our "transmit" EPs/ctxs and vice-versa.
A while ago I discovered a seeming limitation in the tcp;ofi_rxm
provider: namely that it (quoting from the commit) "expects that when an
endpoint or context which seems to be capable of receive operations is
enabled, then a CQ set up to handle those will have been bound to it.
Failing to do that causes us to get an error return when we enable our
tx endpoint(s)".  At the time what I did about this was to turn off all
the rx_attr->caps bits in the fabric domain while creating the transmit
contexts.  I was a little nervous about doing this, since that's info
shared with the provider and it may well have had its own copy.  So now
I'm undoing that workaround and replacing it with something I expect is
a lot more principled, to wit: I'm flagging the CQs (or counters) bound
with both the transmit and receive endpoints/contexts so they can handle
both transmit and receive traffic, independent of what they're actually
expecting.  And I'm leaving the *attr->caps bits the way I found them!