Update recent anonymousCountedRange.chpl test to work with --fast
When adding an error case to anonymousCountedRange.chpl in my last
change to it, I failed to anticipate that --fast testing would skip
the maximal range overflow checking and actually try to iterate over
the maximal range, resulting in a timeout.  Here, I'm breaking the
error case back out into a test of its own that is skipped for --fast
and --no-bounds-checking flags are on.